Movie Library Management System
The program is implemented in C# language with a connection to an SQL database, utilizing 'class' implementation and 'inheritance' structures. It offers options for users to register as either premium or standard users. After registration, users can log in using their usernames and passwords, allowing them access to the film library system. Users can view detailed information about the films in the library. Additionally, administrators have the privilege to log in, manage film information, edit existing films, and add new films to the library.